requirements.txt
python-dotenv==1.0.0
langchain==0.1.0
openai==1.12.0
pydantic==2.6.1
anthropic==0.18.1
huggingface_hub>=0.19.0
requests>=2.28.0 